food_delivery/requirements.txt
asgiref==3.8.1
autopep8==2.3.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
Django==5.2
gunicorn==23.0.0
idna==3.10
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
pycodestyle==2.13.0
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.1.0
python-social-auth==0.3.6
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
gunicorn
whitenoise
dj-database-url
psycopg2-binary
social-auth-app-django